requirements.txt
numpy == 1.21.6
matplotlib == 3.5.3 
wordcloud == 1.8.2.2
flask == 2.2.3
Flask-Caching == 2.0.2
redis == 4.5.1
werkzeug == 2.2.3
gunicorn == 20.1.0
python-dotenv==1.0.0
Pillow==9.5.0
